fixup! Add test for Marconi script transaction indexer
Create the previous tx where outputs are the scripts. Then create the
second transaction where the inputs are the outputs of the previous
transaction, so I could "spend the scripts".

This is WIP as the test can't possibly work as makeTransactionBody is
a pure function and thus it doesn't have access to the generated
scripts nor the hashes, because the inputs (`TxIn = TxId X TxIx`)
don't include it.

Is it implied that I need to run the local testnet -- so that the
lookup from TxIn to script hash happens?